requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.15.0
urllib3==2.6.3
